refactor: move Language class from models to helpers
It's not exactly an eloquent model.
