+++ app/Helpers/Language.php
@@ -14,7 +14,7 @@ declare(strict_types=1);
  * @license    https://www.gnu.org/licenses/agpl-3.0.en.html/ GNU Affero General Public License v3.0
  */
 
-namespace App\Models;
+namespace App\Helpers;
 
 class Language
 {
