+++ phpstan-baseline.neon
@@ -592,7 +592,7 @@ parameters:
 			message: '#^Parameter \#1 \$view of function view expects view\-string\|null, string given\.$#'
 			identifier: argument.type
 			count: 2
-			path: app/Models/Language.php
+			path: app/Helpers/Language.php
 
 		-
 			message: '#^Class App\\Models\\Rss has PHPDoc tag @property for property \$json_torrent with no value type specified in iterable type array\.$#'
